refactor(dedicated.server): move list page as first architecture level
