+++ packages/manager/apps/dedicated/client/app/dedicated/server/details/server.bandwidth-vrack-order.service.js
