+++ packages/manager/apps/dedicated/client/app/dedicated/server/details/server.constants.js
