+++ packages/manager/apps/dedicated/client/app/dedicated/server/details/server.controller.js
