+++ packages/manager/apps/dedicated/client/app/dedicated/server/details/server.feature-availability.js
