+++ packages/manager/apps/dedicated/client/app/dedicated/server/servers.module.js
@@ -6,6 +6,7 @@ const moduleName = 'ovhManagerDedicatedServerServers';
 angular
   .module(moduleName, [])
   .component('dedicatedServerServers', component)
-  .config(routing);
+  .config(routing)
+  .run(/* @ngTranslationsInject:json ./translations */);
 
 export default moduleName;
