--- packages/manager/apps/dedicated/client/app/dedicated/server/servers/translations/Messages_de_DE.json
@@ -1,9 +1,0 @@
-{
-  "dedicated_servers_title": "Dedicated Server",
-  "dedicated_servers_name": "Name",
-  "dedicated_servers_status": "Status",
-  "dedicated_servers_reverse": "Reverse",
-  "dedicated_servers_localization": "Standort",
-  "dedicated_servers_model": "Instanz-Typ",
-  "dedicated_servers_order": "Dedicated Server bestellen"
-}
