--- packages/manager/apps/dedicated/client/app/dedicated/server/servers/translations/Messages_es_ES.json
@@ -1,9 +1,0 @@
-{
-  "dedicated_servers_title": "Servidores dedicados",
-  "dedicated_servers_name": "Nombre",
-  "dedicated_servers_status": "Estado",
-  "dedicated_servers_reverse": "Registro inverso",
-  "dedicated_servers_localization": "Localización",
-  "dedicated_servers_model": "Modelo",
-  "dedicated_servers_order": "Contratar un servidor dedicado"
-}
