--- packages/manager/apps/dedicated/client/app/dedicated/server/servers/translations/Messages_fi_FI.json
@@ -1,9 +1,0 @@
-{
-  "dedicated_servers_title": "Dedicated servers",
-  "dedicated_servers_name": "Name",
-  "dedicated_servers_status": "Status",
-  "dedicated_servers_reverse": "Reverse DNS",
-  "dedicated_servers_localization": "Location",
-  "dedicated_servers_model": "Model",
-  "dedicated_servers_order": "Order a dedicated server"
-}
