--- packages/manager/apps/dedicated/client/app/dedicated/server/servers/translations/Messages_fr_CA.json
@@ -1,9 +1,0 @@
-{
-  "dedicated_servers_title": "Serveurs dédiés",
-  "dedicated_servers_name": "Nom",
-  "dedicated_servers_status": "Statut",
-  "dedicated_servers_reverse": "Reverse",
-  "dedicated_servers_localization": "Localisation",
-  "dedicated_servers_model": "Modèle",
-  "dedicated_servers_order": "Commander un serveur dédié"
-}
