--- packages/manager/apps/dedicated/client/app/dedicated/server/servers/translations/Messages_it_IT.json
@@ -1,9 +1,0 @@
-{
-  "dedicated_servers_title": "Server dedicati",
-  "dedicated_servers_name": "Nome",
-  "dedicated_servers_status": "Stato",
-  "dedicated_servers_reverse": "Reverse",
-  "dedicated_servers_localization": "Localizzazione",
-  "dedicated_servers_model": "Modello",
-  "dedicated_servers_order": "Ordina un server dedicato"
-}
