--- packages/manager/apps/dedicated/client/app/dedicated/server/servers/translations/Messages_pl_PL.json
@@ -1,9 +1,0 @@
-{
-  "dedicated_servers_title": "Serwery dedykowane",
-  "dedicated_servers_name": "Nazwa",
-  "dedicated_servers_status": "Status",
-  "dedicated_servers_reverse": "Rewers",
-  "dedicated_servers_localization": "Lokalizacja",
-  "dedicated_servers_model": "Model",
-  "dedicated_servers_order": "Zamów serwer dedykowany"
-}
