--- packages/manager/apps/dedicated/client/app/dedicated/server/servers/translations/Messages_pt_PT.json
@@ -1,9 +1,0 @@
-{
-  "dedicated_servers_title": "Servidores dedicados",
-  "dedicated_servers_name": "Nome",
-  "dedicated_servers_status": "Estado",
-  "dedicated_servers_reverse": "Reverse",
-  "dedicated_servers_localization": "Localização",
-  "dedicated_servers_model": "Modelo",
-  "dedicated_servers_order": "Encomendar um servidor dedicado"
-}
